Moved major pieces of logic into their own class files. Further expansion will be needed but this version is still functional.
